feat: Add English resume root route
